Moved API datasource to api module for better error handling
