+++ api/src/main/kotlin/fobo66/valiutchik/api/CurrencyRatesDataSource.kt
@@ -14,9 +14,7 @@
  *    limitations under the License.
  */
 
-package fobo66.valiutchik.core.model.datasource
-
-import fobo66.valiutchik.api.Currency
+package fobo66.valiutchik.api
 
 interface CurrencyRatesDataSource {
   suspend fun loadExchangeRates(city: String): Set<Currency>
